working on removing ia dependency
